Created bank_users.main_form.signin_page.forgot_password; Edited bank_users.main_form.main_header and bank_users.main_form.signin_page; Deleted bank_users.main_form.forgot_password
