+++ client_code/bank_users/main_form/main_header/__init__.py
@@ -38,7 +38,7 @@ class main_header(main_headerTemplate):
 
   def login_signup_button_click(self, **event_args):
     """This method is called when the button is clicked"""
-    open_form('bank_users.main_form.login_page')
+    open_form('bank_users.main_form.signin_page')
 
   def button_10_copy_click(self, **event_args):
     """This method is called when the button is clicked"""
