+++ client_code/bank_users/main_form/signin_page/forgot_password/form_template.yaml
@@ -89,9 +89,10 @@ components:
         layout_properties: {grid_position: 'JHGYMR,PKGYXY'}
         event_bindings: {}
       - name: label_2
-        properties: {icon: 'fa:lock', font_size: 20, visible: false}
+        properties: {icon: 'fa:info', font_size: 20, visible: false}
         type: Label
         layout_properties: {grid_position: 'JHGYMR,EBNTMA'}
+        data_bindings: []
       - name: label_5_copy
         properties: {text: Retype Password, bold: true, visible: false}
         type: Label
