+++ client_code/bank_users/main_form/signin_page/form_template.yaml
@@ -73,7 +73,7 @@ components:
       type: Image
       layout_properties: {grid_position: 'LOUJDS,RTKSOW IYLGEA,XVBVAK'}
   - name: column_panel_1
-    properties: {background: '#094586', col_widths: '{"LLBEDS":40,"FZHHNU":5,"TSUFAE":15}'}
+    properties: {background: '#094586', col_widths: '{"LLBEDS":45,"FZHHNU":5,"TSUFAE":10}'}
     type: ColumnPanel
     layout_properties: {grid_position: 'AUEKVR,AGIDMN UVWPRW,XZEJHL', full_width_row: false}
     components:
